requirements.txt
pyobjc-core>=5.2
pyobjc-framework-Cocoa>=5.2
selenium==4.6.0
requests==2.31.0
python-dotenv==1.0.1
coloredlogs==15.0
